requirements.txt
pytest-playwright==0.6.2
pylint==3.3.4
requests==2.32.3